--- a/x/profiles/types/account_test.go
+++ b/x/profiles/types/account_test.go
@@ -376,6 +376,8 @@ func BenchmarkProfile_Update(b *testing.B) {
 		),
 	)
 
+	b.ResetTimer()
+
 	for i := 0; i < b.N; i++ {
 		_, _ = profile.Update(update)
 	}