@RiccardoM 2021-07-23
Re-checking this, to be honest I'm not sure whether we should have this removed inside the `Update` function or not. Ideally, that function can be used by third party developers as well, and those checks are very quick to perform. For this reason, I don't think there's the necessity to remove it at all. We could simply leave it here. Before jumping to the conclusions, though, is it possible to have some data? Like, could we measure the impact on performances that this double check has? I think this should be farily easy to do and it could give us a nice idea whether it is worth or not removing the check.