@bragaz 2021-07-26
@RiccardoM I've performed a bench test on `Update()` and these are the results:

with validation:
```
goos: darwin
goarch: amd64
pkg: github.com/desmos-labs/desmos/x/profiles/types
BenchmarkUpdate

BenchmarkUpdate-1   	  296853	      3937 ns/op
BenchmarkUpdate-2   	  310422	              3490 ns/op
BenchmarkUpdate-3   	  308560	      3546 ns/op
BenchmarkUpdate-4   	  311792	              3612 ns/op
BenchmarkUpdate-5   	  312789	              3666 ns/op
BenchmarkUpdate-6   	  286815	              3789 ns/op
BenchmarkUpdate-7   	  293988	      4159 ns/op
BenchmarkUpdate-8   	  311395	              3671 ns/op
BenchmarkUpdate-9   	  313987	              3518 ns/op
BenchmarkUpdate-10     320106	              3681 ns/op
```

without validation:
```
goos: darwin
goarch: amd64
pkg: github.com/desmos-labs/desmos/x/profiles/types
BenchmarkUpdate

BenchmarkUpdate-1   	 3203461	       380 ns/op
BenchmarkUpdate-2   	 3068451	       378 ns/op
BenchmarkUpdate-3   	 3199720	       367 ns/op
BenchmarkUpdate-4   	 3229094	       381 ns/op
BenchmarkUpdate-5   	 3226158	       372 ns/op
BenchmarkUpdate-6   	 3105884	       386 ns/op
BenchmarkUpdate-7   	 3091993	       370 ns/op
BenchmarkUpdate-8   	 3018584	       386 ns/op
BenchmarkUpdate-9   	 3113236	       374 ns/op
BenchmarkUpdate-10    3172558	       376 ns/op
```
Thoughts?