@RiccardoM 2021-07-27
@bragaz Can you please push the bechmark test code? It's hard to understand what those results mean without having a look at the code