@RiccardoM 2021-07-29
@bragaz After watching at the benchmarks and the code, I think it's safe to delete he `Validate` from within the `Update` method. Could you please update the `Update` documentation telling that it does not validate the updated profile, and this should be done before storing the profile itself? Just as a precaution for external developers